backend/requirements.txt
alembic==1.14.1
bidict==0.23.1
click==8.1.8
dnspython==2.6.1
eventlet==0.33.3
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
h11==0.16.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
MarkupSafe==2.1.5
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.14.3
simple-websocket==1.1.0
six==1.17.0
sqlalchemy==2.0.44
typing-extensions==4.13.2
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.20.2
